Modularise
Moved and mordularised the card page as to have the components and function ready for other pages
